Major refactored of the code to move everything related to receiving data from the OpenBCI Board to its own class for clarity of functions
